admin/requirements.txt
gradio>=4.0.0
pandas>=1.5.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
